full SSL + supervised training is running.
Currently implementing linear probing in order to select the best SSL model to go forth with in the downstream task

First attempt was a normal linear probe,
Now we actually try to add a small DETR head and train that. Not finished with this. Considerations: This will probably takes too long time to train?